Supporting Python3.5 instead of 2.7. Refactored dsp a little.
