--- brewery/brewing/__init__.py
@@ -1,1 +1,0 @@
-from brewhouse import Brewhouse
